+++ verify/unit_test/graph/bfs01.test.cpp
@@ -5,7 +5,7 @@
 #include "../../../src/graph/bfs01.hpp"
 #include "../../../src/graph/dijkstra.hpp"
 void undirected() {
-    int n = rng(1, 200000), m = rng(1, min(200000ll, 1ll * n * (n - 1) / 2));
+    int n = rng(1, 200000), m = rng(1, (int)min(200000ll, 1ll * n * (n - 1) / 2));
     auto [u, v, w] = rng.weighted_graph(n, m, 0, 1, false);
     int weight = rng(1, 1000);
     Graph<int> g(n);
@@ -18,7 +18,7 @@ void undirected() {
     }
 }
 void directed() {
-    int n = rng(1, 200000), m = rng(1, min(200000ll, 1ll * n * (n - 1) / 2));
+    int n = rng(1, 200000), m = rng(1, (int)min(200000ll, 1ll * n * (n - 1) / 2));
     auto [u, v, w] = rng.weighted_graph(n, m, 0, 1, false);
     int weight = rng(1, 1000);
     Graph<int> g(n);
